Home Page Rough
This is a very rough home page outline. I put some basic background colors and fonts for now until we decide on branding. I also am in the process of editing the photos. I added a photo slide show to the top but we can change it to the news.