New Index
So I have added an "index2.html" which is a new index I have made since the old one is being stupid. I HAVE NOT deleted the old index. Also, the new one is responsive. I need to edit the nav bar to the one we like and change out some of the images.