Updates
I updated the manuals page to have the PDFs in it. (Hence all the PDFs.)
The members page now has the link to the FTP server that Chris sent us.
The public page is basically the members page with the bylaws and FTP server taken out and more "learn about VBAS" and join us stuff added in.